Reorganise catalogue into submodules
