+++ tmbr/Sources/App/Modules/Catalogue/Books/Commands/Command/FetchBookCommand.swift
