+++ tmbr/Sources/App/Modules/Catalogue/Books/Models/Book.swift
