+++ tmbr/Sources/App/Modules/Catalogue/Books/Routes/API/Responses/BookResponse.swift
