+++ tmbr/Sources/App/Modules/Catalogue/Catalogue/Commands/Command/Command+SearchCatalogue.swift
